requirements.txt
streamlit>=1.28.0
openai-whisper>=20231106
transformers>=4.30.2
torch>=2.0.1
torchaudio>=2.0.2
ffmpeg-python>=0.2.0
soundfile>=0.12.1
moviepy>=1.0.3